pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "optiprofiler-agent"
version = "0.1.0dev0"
description = "AI Agent system for OptiProfiler — product advisor, auto-debug, and results interpretation"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
authors = [
    {name = "OptHuang"},
]
keywords = ["optimization", "benchmarking", "solver", "agent", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Topic :: Scientific/Engineering :: Mathematics",
]

dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "langgraph>=0.4",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
rag = [
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
]
interpret = [
    "PyMuPDF>=1.24",
]
anthropic = [
    "langchain-anthropic>=0.3",
]
all = [
    "optiprofiler-agent[rag,interpret,anthropic]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
optiprofiler-agent = "optiprofiler_agent.cli:main"

[project.urls]
Homepage = "https://www.optprof.com"
Repository = "https://github.com/optiprofiler/optiprofiler-agent"

[tool.setuptools.packages.find]
include = ["optiprofiler_agent*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
